🚚 move: EvalLogSearchModal 폴더 분리
