+++ app/src/pages/EvalLogSearchPage/index.tsx
@@ -10,7 +10,7 @@ import { useInfiniteScroll } from '@utils/useInfiniteScroll';
 import { isEqual } from 'lodash-es';
 import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { EvalLogSearchModal } from './EvalLogSearchModal';
+import { EvalLogSearchModal } from './EvalLogSearchModal/EvalLogSearchModal';
 import { EvalLogSearchResult } from './EvalLogSearchResult';
 import { EvalLogSearchTitle } from './EvalLogSearchTitle';
